+++ packages/core/src/utils/providers/index.ts
@@ -1,2 +1,3 @@
 export * from './blockfrost.provider'
+export * from './demeter.provider'
 export * from './ogmios.provider'